XMLHttpRequest/open-sync-open-send.htm

<!DOCTYPE html>
<html>
  <head>
    <title>XMLHttpRequest: open() (sync) - send() - open()</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      var test = async_test()
      test.step(function() {
        var client = new XMLHttpRequest(),
            result = [],
            expected = [1,4,1]
        client.onreadystatechange = function() {
          test.step(function() {
            result.push(client.readyState)
          })
        }
        client.open("GET", "folder.txt")
        client.send(null)
        client.open("GET", "folder.txt", false)
        assert_array_equals(result, expected)
        assert_equals(client.responseXML, null)
        assert_equals(client.responseText, "")
        assert_equals(client.status, 0)
        assert_equals(client.statusText, "")
        assert_equals(client.getAllResponseHeaders(), "")
        test.done()
      })
    </script>
  </body>
</html>
